@@ -1,30 +1,28 @@
 name: huldra
 description: A sample command-line application.
-version: 0.9.3
+version: 0.9.4
 # homepage: https://www.example.com
 environment:
-  sdk: ">=2.10.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
-  crypto: ^2.1.4
-  hive: ^1.4.4+1
-  injector: ^1.0.8
-  json_serializable: ^3.5.0
-  moor: ^3.4.0
-  nyxx: ^0.30.0
-  path: ^1.6.0
-  sqlite3: ^0.1.7
-  yaml_config: ^0.1.2
+  crypto: ^3.0.1
+  injector: ^2.0.0
+  json_serializable: ^4.1.2
+  moor: ^4.3.0
+  nyxx: ^1.0.2
+  path: ^1.8.0
+  sqlite3: ^1.1.1
+  yaml: ^3.1.0
 
 dev_dependencies:
-  build_runner: ^1.10.4
-  hive_generator: ^0.8.1
-  moor_generator: ^3.4.0
-  pedantic: ^1.8.0
-  test: ^1.6.0
+  build_runner: ^2.0.3
+  moor_generator: ^4.3.0
+  pedantic: ^1.11.0
+  test: ^1.17.4
 
-dependency_overrides:
-  nyxx:
-    git:
-      url: https://github.com/l7ssha/nyxx.git
-      ref: development
+# dependency_overrides:
+#   nyxx:
+#     git:
+#       url: https://github.com/l7ssha/nyxx.git
+#       ref: development
